test(gui): count the gate needles, and split the hiding scan out
`GreyOutParityTests.gatedEditors` asserted with a bare
`contains`, so a file gating TWO rows with one expression stayed
green with either of the pair deleted. That was not theoretical:
the App Bar's Gap-indicator gate guards Highlight and Active item
with an identical call, and deleting only Highlight's — leaving
the swatch editable while nothing it tints is drawn — passed.
`GreyOutAnchorTests` had already solved this with a `count:`
field and named the hazard in its own docstring; this is the same
field, and the same reasoning, one file over.

Re-proved by deleting one of the pair: 1 of 2, red.

The hide-instead-of-dim scan moves to `GreyOutHidingTests` — the
count field pushed the file past the 350-line ceiling, and §5
says split before that, not after.

Also states plainly that the `BarColorCards` exemption needle
watches a SHAPE and not behaviour: no Advanced Colours row is
exempt today, so its green is not coverage. It is there so
whoever adds the first exemption finds the renderer already
honoring the flag rather than silently ignoring it.

Five catalog corrections ride along, all from the localization
audit of the strings the previous commit added:

- fr `sidebar.colors` takes "Couleurs & anim." (owner's call).
  No French rendering of "Colors & Motion" fits the sidebar's
  ~120 pt budget — the literal is 141 pt — and the alternative
  was dropping the Motion half of the name in French alone.
- The four CJK renderings of `colors.border_off.help` named the
  wrong thing. The Japanese one called the control
  「フォーカスの境界線」 while the switch it sends the user to
  ships as 「フォーカスボーダーを有効化」 — a popover naming a
  control its own destination does not have. All four also left
  WHICH two of the card's three swatches go dark to a pronoun,
  which cannot carry it; they now name the pair using each
  locale's own swatch labels, so popover and swatches share
  their words.
