+++ Sources/KiwiDeskCore/Resources/Locales/fr.json
@@ -710,7 +710,7 @@
   "sidebar.app_rules": "Règles d’apps",
   "sidebar.bars": "Barres",
   "sidebar.behavior": "Comportement",
-  "sidebar.colors": "Couleurs & motion",
+  "sidebar.colors": "Couleurs & anim.",
   "sidebar.gaps_borders": "Écarts & bordures",
   "sidebar.general": "Général",
   "sidebar.layout": "Dispositions par défaut",
