+++ Sources/KiwiDeskCore/Resources/Locales/zh-Hant.json
@@ -162,7 +162,7 @@
   "color_field.swatch.help_auto": "自動 — 依循淺色與深色模式",
   "color_panel.done": "完成",
   "colors.app_bar_off.help": "沒有任何佈局顯示 App Bar，因此不會繪製它的顏色。請在「資訊列」中開啟一個。",
-  "colors.border_off.help": "焦點邊框已關閉，因此不會繪製它的兩種顏色。請在「間距與邊框」中開啟它。",
+  "colors.border_off.help": "焦點邊框已關閉，因此不會繪製焦點視窗與非焦點視窗的顏色。請在「間距與邊框」中開啟它。",
   "colors.borders.caption": "焦點視窗周圍的邊框，以及 Sticky 視窗上的標記。",
   "colors.borders.title": "邊框顏色",
   "colors.drag.caption": "你拿起的視窗，以及它將放入的位置。",
